Move static methods to utils and add tests
